pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cloudinary>=1.44.1",
    "cohere>=5.15.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.16.0",
]
